requirements.txt
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.17.1
djangorestframework-simplejwt==5.5.1
firebase-admin==7.4.0
gunicorn==25.3.0
dj-database-url==3.1.2
python-decouple==3.8
whitenoise==6.12.0
psycopg2-binary==2.9.11
redis==7.4.0
EOF 
